requirements.txt
Pillow
onnxruntime
rembg
ttkbootstrap
